pyproject.toml
[tool.poetry]
name = "giant-plugins"
version = "1.0.0"
description = "Adds a generic list of plugins for use within projects"
authors = ["Will-Hoey <will.hoey@giantdigital.co.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/giantmade/giant-plugins"
repository = "https://github.com/giantmade/giant-plugins"
keywords = ["plugins", "app"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]
packages = [
    { include = "giant_plugins" }
]

[tool.poetry.dependencies]
python = "^3.6"
django-filer = "*"
giant-mixins = "*"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
django-cms = "^3.7.2"
django = "2.2"
pytest-cov = "^2.10.0"
pytest-django = "^3.9.0"
six = "^1.15.0"
django-polymorphic = "3.0.0"

[[tool.poetry.source]]
name = "TestPyPi"
url = "https://test.pypi.org/simple/"
secondary = true

[tool.black]
line-length = 99
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.ipython
  | \.mypy_cache
  | \.ropeproject
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"